Issue #640: define abstract unet interface
